% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mocaclock.R
\name{calculate_mocaclock}
\alias{calculate_mocaclock}
\title{Calculate MoCA clock drawing test total}
\usage{
calculate_mocaclock(MOCACLOC, MOCACLON, MOCACLOH)
}
\arguments{
\item{MOCACLOC}{values for clock contour}

\item{MOCACLON}{values for clock numbers}

\item{MOCACLOH}{values for clock hands}
}
\description{
Sum of MOCACLOC, MOCACLON, and MOCACLOH, but applies \code{\link[=valid_values_only]{valid_values_only()}} before
adding values.
}
