% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reytotal.R
\name{calculate_reytotal}
\alias{calculate_reytotal}
\title{Calculate Rey AVLT Total score}
\usage{
calculate_reytotal(REY1REC, REY2REC, REY3REC, REY4REC, REY5REC)
}
\arguments{
\item{REY1REC}{values for RAVLT Trial 1 total recall}

\item{REY2REC}{values for RAVLT Trial 2 total recall}

\item{REY3REC}{values for RAVLT Trial 3 total recall}

\item{REY4REC}{values for RAVLT Trial 4 total recall}

\item{REY5REC}{values for RAVLT Trial 5 total recall}
}
\description{
Sum of REY1REC, ..., REY5REC, but applies \code{\link[=valid_values_only]{valid_values_only()}} before
adding values. This removes any values not between 0 and 15.
}
