% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_methods.R
\docType{data}
\name{default_methods}
\alias{default_methods}
\title{Default Standardization Methods}
\format{
A named list. Each entry is named for a neuropsych score, and contains
a named vector with two entires: \code{method} gives the default method, and
\code{version} the default version.
}
\usage{
default_methods
}
\description{
Default Standardization Methods
}
\keyword{datasets}
