% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age_group.R
\name{get_age_group}
\alias{get_age_group}
\title{Get Age Group for Standardized Scores}
\usage{
get_age_group(age)
}
\arguments{
\item{age}{numeric vector with ages to be converted to groups}
}
\description{
Get various versions of age groups for finding standardized scores.
}
