% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_educ_group.R
\name{get_educ_group}
\alias{get_educ_group}
\title{Get education group}
\usage{
get_educ_group(education)
}
\arguments{
\item{education}{A numeric vector representing years of education.}
}
\value{
A factor representing education groups, with levels \code{"<13"}, \code{"13-15"}, \code{"16"}, and \code{">16"}.
}
\description{
Get education group
}
