% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_race_groups.R
\name{get_race_group}
\alias{get_race_group}
\title{Get Race Group for Standardized Scores}
\usage{
get_race_group(race)
}
\arguments{
\item{race}{numeric vector with race (according to NACC; see \code{rdd$RACE} for details) to be converted to groups}
}
\description{
Get race group for finding standardized scores.
}
