% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_score.R
\name{get_ss_score}
\alias{get_ss_score}
\title{Get Scaled Scores for use in T-score cacluations}
\usage{
get_ss_score(raw_score, var_name)
}
\arguments{
\item{raw_score}{numeric vector with raw scores}

\item{var_name}{string indicating what variable}
}
\description{
Helper function to look up unadjusted scaled scores in the \code{\link{ss}} list
for use in the fully adjusted T-scores.
}
