% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implemented_methods_functions.R
\name{implemented_std_methods}
\alias{implemented_std_methods}
\alias{std_methods_implemented_for}
\title{Implemented Methods for Specific Variable}
\usage{
implemented_std_methods(var_name)

std_methods_implemented_for(method, version)
}
\arguments{
\item{var_name}{a NACC variable to look up implemented methods for.}

\item{method}{method to look up. One of \code{norms}, \code{regression}, or \code{T-score}}

\item{version}{version of method to look up. Either \code{nacc} (indicating the use
of published methods) or \code{updated} (indicating the use of updated versions
of the published methods, using data available as of June 2024). Ignored if
\code{method} is \code{T-score}.}
}
\description{
\code{implemented_std_methods} gives a table that specifies which method/version
is implemented for \code{var_name}

\code{std_methods_implemented_for} gives the variables that can be standardized
using \code{method}/\code{version}.
}
