% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_from_std_data.R
\name{methods_from_std_data}
\alias{methods_from_std_data}
\title{Helper to extract methods from standardized columns}
\usage{
methods_from_std_data(std_data, std_cols)
}
\arguments{
\item{std_data}{\code{data.frame}/\code{tibble} with some standardized columns}

\item{std_cols}{optional; specify columns to get methods for. If omitted, will
return method attribute from all columns that have said attribute}
}
\description{
Helper to extract methods from standardized columns
}
