% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mns_to_nacc.R
\docType{data}
\name{mns_to_nacc}
\alias{mns_to_nacc}
\title{Translate NACC variable names to Mayo Normative Study variable names}
\format{
A named vector where names of entries correspond to Mayo Normative
Study variable names, and entries contain NACC variable names.
}
\usage{
mns_to_nacc
}
\description{
Translate NACC variable names to Mayo Normative Study variable names
}
\keyword{datasets}
