% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nacc_to_mns.R
\docType{data}
\name{nacc_to_mns}
\alias{nacc_to_mns}
\title{Translate Mayo Normative Study variable names to NACC variable names}
\format{
A named vector where names of entries correspond to NACC variable
names, and entries contain Mayo Normative Study variable names.
}
\usage{
nacc_to_mns
}
\description{
Translate Mayo Normative Study variable names to NACC variable names
}
\keyword{datasets}
