% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NpsychBatteryNorms-package.R
\docType{package}
\name{NpsychBatteryNorms-package}
\alias{NpsychBatteryNorms}
\alias{NpsychBatteryNorms-package}
\title{NpsychBatteryNorms: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rtrane@wisc.edu}

}
\keyword{internal}
