% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_scores.R
\name{std_scores}
\alias{std_scores}
\title{Standardize Raw Cognitive Scores to Z-Scores}
\usage{
std_scores(
  raw_scores,
  var_name,
  education,
  age,
  sex = c("m", "f"),
  delay = NULL,
  method = c("norms", "regression", "T-score"),
  version = c("nacc", "updated"),
  print_messages = T
)
}
\arguments{
\item{raw_scores}{Numeric vector of raw scores}

\item{var_name}{String with name of the variable. Used to get correct means and standard deviations, or regression coefficients.}

\item{education}{Numeric vector with years of education. Used for subset of variables only.}

\item{age}{Numeric vector with ages in years}

\item{sex}{Character vector with sex of participants. Must be either "m" (for male) or "f" (for female). Used for subset of variables only}

\item{delay}{Only for standardizing \code{MEMUNITS} ("Logical Memory, Delayed") in which case it is the time of delay in minutes}

\item{method}{String specifying method to use for standardization.}

\item{version}{String specifying version to use. "nacc" = 2020 numbers
published by NACC. "updated" = numbers calculated using data freeze from
June of 2024.}

\item{print_messages}{Logical (default: TRUE); should messages be printed?}
}
\description{
Calculate z-scores as \verb{(raw score - average) / sd}. Available for a wide array
of cognitive scores both using normative data, and regression based models. A
wrapper around \code{std_scores_using_norms()}, \code{std_scores_using_regression()} and
\code{std_scores_using_T()}.
}
\examples{
std_scores(
  raw_scores = c(21, 15),
  var_name = "MOCATOTS",
  sex = c("m", "f"),
  age = c(57, 70),
  education = c(15, 13),
  method = "regression",
  version = "nacc"
)

}
