% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_scores_using_norms.R
\name{std_scores_using_norms}
\alias{std_scores_using_norms}
\title{Standardized Scores using Norms}
\usage{
std_scores_using_norms(raw_scores, var_name, education, age, sex, m_sd)
}
\arguments{
\item{raw_scores}{Numeric vector of raw scores}

\item{var_name}{String with name of the variable. Used to get correct means and standard deviations, or regression coefficients.}

\item{education}{Numeric vector with years of education. Used for subset of variables only.}

\item{age}{Numeric vector with ages in years}

\item{sex}{Character vector with sex of participants. Must be either "m" (for male) or "f" (for female). Used for subset of variables only}

\item{m_sd}{data.frame or tibble with means and standard deviations to use for
standardizing the raw scores. Should also include columns for age group,
education group, and sex (if desired). For examples, see \link{normative_scores_2024},
\link{normative_scores_2020}, and \link{ravlt_trials_m_sd}.}
}
\description{
Returns standardized scores by calculating \code{(raw_score - mean) / sd}. Means
and standard deviations are given in the \code{m_sd} argument, and must be dependent
on age, with the option to also depend on sex, and education.
}
