% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_scores_using_regression.R
\name{std_scores_using_regression}
\alias{std_scores_using_regression}
\title{Standardize Raw Cognitive Scores using Regression Models}
\usage{
std_scores_using_regression(
  raw_scores,
  var_name,
  reg_coefs,
  age,
  education,
  sex,
  race,
  delay,
  sd
)
}
\arguments{
\item{raw_scores}{Numeric vector of raw scores}

\item{var_name}{String with name of the variable. Used to get correct means and standard deviations, or regression coefficients.}

\item{reg_coefs}{named vector with coefficients to be used for calculating
expected score. Can include the following named entries:
\itemize{
\item \code{intercept}
\item \code{age}
\item \code{education}
\item \code{sex}
\item \code{race}
\item \code{delay}
}}

\item{age}{Numeric vector with ages in years}

\item{education}{Numeric vector with years of education. Used for subset of variables only.}

\item{sex}{Character vector with sex of participants. Must be either "m" (for male) or "f" (for female). Used for subset of variables only}

\item{race}{Character vector with race of participants. Must be either "White" or "Other". Used only for updated regression models.}

\item{delay}{Only for standardizing \code{MEMUNITS} ("Logical Memory, Delayed") in which case it is the time of delay in minutes}

\item{sd}{estimated standard deviation}
}
\description{
Get normative scores by standardizing using regression models adjusting for
one or all of age, sex and education.
}
\section{Normative Scores from Regression Model}{
In general, we standardize scores by subtracting an estimate of the (possibly
conditional) expected value of the variable considered, and divide the
difference by an estimate of the (possibly conditional) standard deviation.

When we aim to do this using a regression model that adjusts for age, sex,
and years of education, we find the conditional expected value as

\deqn{
\mu = \beta_0 + \beta_{age} \cdot \verb|age| +
\beta_{education} \cdot \verb|education| +
\beta_{sex} \cdot \verb|sex|
}

where \eqn{\verb|sex|} is \eqn{1} if the patient is female, and \eqn{0}
otherwise. \code{std_scores_using_regression} takes the coefficients in a named
vector \eqn{\verb|c(intercept = |\beta_0\verb|, age = |\beta_{age}\verb|, education = |\beta_{education}\verb|, sex = |\beta_{sex}\verb|)|}.

The normative score is then given as \eqn{\verb|(raw_scores - |\mu\verb|)/sd|},
where \code{sd} is the estimated standard deviation passed to the \code{sd} argument.
This will typically be the root mean squared error (RMSE) from the linear
regression model.

One final caveat: for consistency, we "flip the sign" for normative scores
where higher raw score values correspond to worse performance. This is the
case for the trail making scores (both written and oral) related to time and
number of errors.
}

