% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_scores_using_T.R
\name{std_scores_using_T}
\alias{std_scores_using_T}
\title{Standardized Scores using T-scores}
\usage{
std_scores_using_T(raw_scores, var_name, age, education, sex)
}
\arguments{
\item{raw_scores}{Numeric vector of raw scores}

\item{var_name}{String with name of the variable. Used to get correct means and standard deviations, or regression coefficients.}

\item{age}{Numeric vector with ages in years}

\item{education}{Numeric vector with years of education. Used for subset of variables only.}

\item{sex}{Character vector with sex of participants. Must be either "m" (for male) or "f" (for female). Used for subset of variables only}
}
\description{
Standardized Scores using T-scores
}
