% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_score_coefs.R
\docType{data}
\name{t_score_coefs}
\alias{t_score_coefs}
\title{Regression Coefficients for T-score Based Standardization}
\format{
A named list with one entry for each NACC variable for which T-scores
have been implemented. Each entry contains the coefficients needed for the
corresponding model. This is a subset of \code{intercept}, \code{age}, \code{age^2}, \code{sex},
\code{education}, \code{intercept2}, \code{sd}, \code{intercept3}, \code{education^0.5}, \verb{age^2_2}.
}
\usage{
t_score_coefs
}
\description{
Regression coefficients used to obtain standardized scores using T-scores.
}
\references{
The coefficients implemented here are published in the following two
articles:
\itemize{
\item \insertRef{strickerMayoNormativeStudies2021}{NpsychBatteryNorms}
\item \insertRef{karstensMayoNormativeStudies2023}{NpsychBatteryNorms}
}
}
\keyword{datasets}
