% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_values_only.R
\name{valid_values_only}
\alias{valid_values_only}
\title{Replace values with labels}
\usage{
valid_values_only(raw_score, var_name = "ANIMALS", remove_errorcodes = FALSE)
}
\arguments{
\item{raw_score}{numeric vector of raw scores.}

\item{var_name}{character specifying what test the raw scores are from. Must
be one of the names of the entries in \code{rdd}}

\item{remove_errorcodes}{should values listed in the \code{rdd$var_name$codes}
be replaced by \code{NA} in output? Either FALSE (default; do not remove valid
codes), TRUE (remove all codes), or a numeric vector with codes to remove.}
}
\description{
Replace values with labels
}
