% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_to_labels.R
\name{values_to_labels}
\alias{values_to_labels}
\title{Replace codes with labels}
\usage{
values_to_labels(raw_score, var_name = "SEX")
}
\arguments{
\item{raw_score}{numeric vector of raw scores.}

\item{var_name}{character specifying what test the raw scores are from. Must
be one of the names of the entries in \code{rdd}}
}
\description{
Replace codes with labels
}
